refactor: reorganize vector-stores/github module structure
- Move business logic from app/api to lib/vector-stores/github
- Split responsibilities into clear subdirectories:
  - ingest/: Repository ingestion logic
  - query/: Query service implementation
  - Common files (types.ts, database.ts) at root level
- Rename functions for clarity:
  - fetchTargetGitHubRepositories → fetchIngestTargets
  - github-blobs.ts → ingest-github-blobs.ts
- Extract utilities into dedicated files:
  - github-client.ts: Octokit client creation
  - fetch-ingest-targets.ts: CRON target selection
  - process-repository.ts: Main ingestion logic with status updates
- Remove redundant index.ts files within subdirectories

This improves code organization, reusability, and maintainability.
