+++ apps/studio.giselles.ai/app/api/vector-stores/github/ingest/route.ts
@@ -1,17 +1,8 @@
-import { fetchDefaultBranchHead } from "@giselle-sdk/github-tool";
-import { DocumentLoaderError, RagError } from "@giselle-sdk/rag";
-import { captureException } from "@sentry/nextjs";
 import type { NextRequest } from "next/server";
-import { ingestGitHubBlobs } from "./ingest-github-repository";
-import { createIngestTelemetrySettings } from "./telemetry";
-import type { TargetGitHubRepository } from "./types";
 import {
-	buildOctokit,
-	fetchTargetGitHubRepositories,
-	updateRepositoryStatusToCompleted,
-	updateRepositoryStatusToFailed,
-	updateRepositoryStatusToRunning,
-} from "./utils";
+	fetchIngestTargets,
+	processRepository,
+} from "@/lib/vector-stores/github";
 
 export const maxDuration = 800;
 
@@ -23,113 +14,9 @@ export async function GET(request: NextRequest) {
 		});
 	}
 
-	const targetGitHubRepositories = await fetchTargetGitHubRepositories();
+	const targetGitHubRepositories = await fetchIngestTargets();
 
 	await Promise.all(targetGitHubRepositories.map(processRepository));
 
 	return new Response("ok", { status: 200 });
-}
-
-async function processRepository(
-	targetGitHubRepository: TargetGitHubRepository,
-) {
-	const { owner, repo, installationId, teamDbId, dbId } =
-		targetGitHubRepository;
-
-	try {
-		await updateRepositoryStatusToRunning(dbId);
-
-		const octokitClient = buildOctokit(installationId);
-		const commit = await fetchDefaultBranchHead(octokitClient, owner, repo);
-		const source = {
-			owner,
-			repo,
-			commitSha: commit.sha,
-		};
-
-		const telemetry = await createIngestTelemetrySettings(
-			teamDbId,
-			`${owner}/${repo}`,
-		);
-
-		await ingestGitHubBlobs({
-			octokitClient,
-			source,
-			teamDbId,
-			telemetry,
-		});
-
-		await updateRepositoryStatusToCompleted(dbId, commit.sha);
-	} catch (error) {
-		console.error(
-			`Failed to ingest GitHub Repository: teamDbId=${teamDbId}, repository=${owner}/${repo}`,
-			error,
-		);
-
-		const { errorCode, retryAfter } = extractErrorInfo(error);
-
-		captureException(error, {
-			extra: {
-				owner,
-				repo,
-				teamDbId,
-				errorCode,
-				retryAfter,
-				errorContext:
-					error instanceof DocumentLoaderError ? error.context : undefined,
-			},
-		});
-
-		await updateRepositoryStatusToFailed(dbId, {
-			errorCode,
-			retryAfter,
-		});
-	}
-}
-
-/**
- * Extract error code and retry time from an error
- * @param error The error to extract information from
- * @returns Error code and retry time
- */
-function extractErrorInfo(error: unknown): {
-	errorCode: string;
-	retryAfter: Date | null;
-} {
-	if (error instanceof DocumentLoaderError) {
-		const errorCode = error.code;
-
-		let retryAfter: Date | null;
-		switch (error.code) {
-			case "DOCUMENT_NOT_FOUND":
-			case "DOCUMENT_TOO_LARGE":
-				// Not retryable
-				retryAfter = null;
-				break;
-			case "DOCUMENT_RATE_LIMITED":
-				retryAfter = error.getRetryAfterDate() ?? new Date();
-				break;
-			case "DOCUMENT_FETCH_ERROR":
-				retryAfter = new Date();
-				break;
-			default: {
-				const _exhaustiveCheck: never = error.code;
-				throw new Error(`Unknown error code: ${_exhaustiveCheck}`);
-			}
-		}
-
-		return { errorCode, retryAfter };
-	}
-
-	if (error instanceof RagError) {
-		return {
-			errorCode: error.code,
-			retryAfter: new Date(),
-		};
-	}
-
-	return {
-		errorCode: "UNKNOWN",
-		retryAfter: new Date(),
-	};
 }
