--- apps/studio.giselles.ai/app/api/vector-stores/github/ingest/types.ts
@@ -1,8 +1,0 @@
-export type TargetGitHubRepository = {
-	dbId: number;
-	owner: string;
-	repo: string;
-	teamDbId: number;
-	installationId: number;
-	lastIngestedCommitSha: string | null;
-};
